+++ src/components/Preferences/Modal.jsx
@@ -1,0 +1,48 @@
+import { motion } from "framer-motion";
+import Backdrop from "./Backdrop";
+import { usePrefStore } from "../../store";
+
+const Modal = ({ handleClose }) => {
+
+    const lang = usePrefStore((state) => state.lang)
+    const setLang = usePrefStore((state) => state.setLang)
+
+    return (
+        <Backdrop onClick={handleClose}>
+            <motion.div
+                onClick={(e) => e.stopPropagation()}
+                className="relative flex flex-col items-center p-6 rounded-lg shadow-lg dark:bg-gray-800 z-60"
+                initial={{ opacity: 0, x: 20 }}
+                animate={{ opacity: 1, x: 0 }}
+                exit={{ opacity: 0, x: -20 }}
+                transition={{ duration: 0.13 }}
+            >
+
+                <div className="w-full mb-3 text-left">
+                    <h1 style={{fontStyle: "oblique"}}>Preferences</h1>
+                </div>
+
+                <div id="settingsWindow" className="p-3">
+                    <div id="setting" className="flex items-center">
+                        <p>Language:</p>
+                        <div className="inline-flex ml-3 rounded-md shadow-xs dark:border-slate-700" role="group">
+                            <button type="button" onClick={() => setLang("en")} className={`px-4 py-2 border-t-3 border-b-3 text-sm dark:border-gray-700 font-medium rounded-s-lg focus:z-10 ${lang === "en" ? "bg-blue-700 text-white dark:bg-blue-600" : "bg-white text-gray-900 hover:bg-gray-100 hover:text-blue-700 dark:bg-gray-800 dark:text-white dark:hover:bg-gray-700"}`}>
+                                English (en)
+                            </button>
+                            <button type="button" onClick={() => setLang("kr")} className={`px-4 py-2 border-t-3 border-b-3 text-sm dark:border-gray-700 font-medium focus:z-10 ${lang === "kr" ? "bg-blue-700 text-white dark:bg-blue-600" : "bg-white text-gray-900 hover:bg-gray-100 hover:text-blue-700 dark:bg-gray-800 dark:text-white dark:hover:bg-gray-700"}`}>
+                                Korean (ko)
+                            </button>
+                            <button type="button" onClick={() => setLang("ja")} className={`px-4 py-2 border-t-3 border-b-3 text-sm dark:border-gray-700 rounded-e-lg font-medium focus:z-10 ${lang === "ja" ? "bg-blue-700 text-white dark:bg-blue-600" : "bg-white text-gray-900 hover:bg-gray-100 hover:text-blue-700 dark:bg-gray-800 dark:text-white dark:hover:bg-gray-700"}`}>
+                                Japanese (ja)
+                            </button>
+                        </div>
+                    </div>
+                </div>
+
+                <button onClick={handleClose} className="p-3 rounded-md dark:bg-slate-400">Close</button>
+            </motion.div>
+        </Backdrop>
+    );
+}
+
+export default Modal;