+++ src/context.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+// Keeps track of if the modal is open or not
+export const PrefModalOpenContext = createContext(false)