+++ src/scripts/GetEpisode.jsx
@@ -14,7 +14,8 @@ export default function GetEpisode() {
                 setLoading(true);
                 setError(null);
 
-                const path = window.location.pathname;
+                let path = window.location.pathname;
+                path = path.slice(3) // slices the first slash then the language code. The language code always consists of 2 characters, and follows the ISO 639-1 standard.
                 const segments = path.split('/').filter(Boolean);
 
                 const storytype = segments[0] ?? '';