+++ CHANGELOG.md
@@ -10,6 +10,20 @@ Date format: `YYYY-MMM-DD`
 ## [Unreleased]
 - Nothing yet
 
+## [0.3.0] - 2025-Feb-10
+### Added
+- Zustand for global state management; used to store preferences
+- User's preferred language now appears on the URL path
+- Preferences can be updated with the new preferences modal
+
+### Fixed
+- Files and scripts which rely on the URL path has been adjusted accordingly to account for the language code
+- Header links have been updated
+
+### Removed
+- Unused script.js
+
+
 ## [0.2.0] - 2025-Feb-09
 ### Added
 - root.jsx has new components NumberGuard, used to check if the URL paths are made up of only numbers. Used in QUN.
@@ -35,5 +49,6 @@ Date format: `YYYY-MMM-DD`
 - Everything!
 
 
+[0.2.0]: https://github.com/lorearchive/LADR/compare/v0.2.0...v0.3.0 
 [0.2.0]: https://github.com/lorearchive/LADR/compare/v0.1.1...v0.2.0 
 [0.1.1]: https://github.com/lorearchive/LADR/compare/v0.1.0...v0.1.1